backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
django-environ
Pillow
psycopg[binary]>=3.1
gunicorn
whitenoise
